+++ src/atoms/index.ts
@@ -10,5 +10,6 @@ export * from './linePager';
 export * from './pill';
 export * from './progressCounterItem';
 export * from './tabBarItem';
+export * from './titledCollapsibleBox';
 export * from './webView';
 export * from './checkbox';