+++ src/theming/themeBuilder.tsx
@@ -14,6 +14,7 @@ import { buildPillThemes } from '../atoms/pill';
 import { buildPrettyTextThemes } from '../atoms/prettyText';
 import { buildProgressCounterItemThemes } from '../atoms/progressCounterItem';
 import { buildTabBarItemThemes } from '../atoms/tabBarItem';
+import { buildTitledCollapsibleBoxThemes } from '../atoms/titledCollapsibleBox';
 import { buildWebViewThemes } from '../atoms/webView';
 import { buildBoxThemes } from '../particles/box';
 import { buildAlternateColors, buildColors } from '../particles/colors';
@@ -54,6 +55,7 @@ export const buildTheme = (inputTheme?: RecursivePartial<ITheme>): ITheme => {
   const linePagerThemes = buildLinePagerThemes(colors, dimensions, boxThemes, inputTheme?.linePagers);
   const progressCounterItemThemes = buildProgressCounterItemThemes(colors, dimensions, textThemes, boxThemes, inputTheme?.progressCounterItems);
   const tabBarItemThemes = buildTabBarItemThemes(colors, dimensions, textThemes, boxThemes, inputTheme?.tabBarItems);
+  const titledCollapsibleBoxThemes = buildTitledCollapsibleBoxThemes(colors, dimensions, textThemes, boxThemes, inputTheme?.titledCollapsibleBoxes);
 
   return merge<ITheme>({
     // Base
@@ -86,6 +88,7 @@ export const buildTheme = (inputTheme?: RecursivePartial<ITheme>): ITheme => {
     linePagers: linePagerThemes,
     progressCounterItems: progressCounterItemThemes,
     tabBarItems: tabBarItemThemes,
+    titledCollapsibleBoxes: titledCollapsibleBoxThemes,
   }, inputTheme, {
     // NOTE(krish): this is here so the font replacement doesn't get overridden
     fonts,