+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ## [Unreleased]
 
 ### Added
+- [MINOR] Created `TitledCollapsibleBox` atom component
+- [MINOR] Created `StatefulTitledCollapsibleBox` atom component
 
 ### Changed